chore: rename utils in kebad-case
