+++ src/components/Modal/Modal.tsx
@@ -7,8 +7,8 @@ import {Portal} from '../Portal';
 import type {DOMProps, QAProps} from '../types';
 import {FocusTrap} from '../utils/FocusTrap';
 import {block} from '../utils/cn';
-import type {LayerCloseReason, LayerExtendableProps} from '../utils/layerManager';
-import {useLayer} from '../utils/layerManager';
+import type {LayerCloseReason, LayerExtendableProps} from '../utils/layer-manager';
+import {useLayer} from '../utils/layer-manager';
 import {getCSSTransitionClassNames} from '../utils/transition';
 import {useRestoreFocus} from '../utils/useRestoreFocus';
