+++ src/components/Popup/Popup.tsx
@@ -7,8 +7,8 @@ import {Portal} from '../Portal';
 import type {DOMProps, QAProps} from '../types';
 import {FocusTrap, useParentFocusTrap} from '../utils/FocusTrap';
 import {block} from '../utils/cn';
-import {useLayer} from '../utils/layerManager';
-import type {LayerExtendableProps} from '../utils/layerManager';
+import {useLayer} from '../utils/layer-manager';
+import type {LayerExtendableProps} from '../utils/layer-manager';
 import {getCSSTransitionClassNames} from '../utils/transition';
 import {usePopper} from '../utils/usePopper';
 import type {PopperAnchorRef, PopperPlacement, PopperProps} from '../utils/usePopper';
